Rule 10: Comments in your SCSS
Basic stuff but worth a mention....
We all have a different way of putting comments in. So many are ugly, and more of a distraction than useful. Also, sometimes I want to comment out a whole block of SCSS, and I can't quickly because of so many /* comments already in there.